refactor(trusty-audit): split the OSV run-index roll-up into its own module
`osv.rs` measured 513 SLOC once tracked, over the 500 production cap. The
roll-up is the natural seam: `osv` answers for ONE repository and writes one
`osv.json`; `osv_rollup` folds every repository's file back off disk into
the count table `index.md` renders. Two questions, two modules, no behaviour
change.

Refs #6780
